레이스 컨디션 해결 및, 기능에 따른 뷰모델 파일 세분화
